Fix IPv6 address returned from QUrl::host
When passing an IPv6 address through QNetworkProxyQuery, it
is stored in a QUrl internally.
There was a bug in QUrl where it strips the [] surrounding an
IPv6 address only if they were present in the input, otherwise
it added them.
Now the behaviour is the same as Qt5 ([] are always stripped).

Change-Id: I42e020ce30d18a4108f1dd4428809fed07991680
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>