Always process window (de)activated events
QApplication::setActiveWindow() is fully able to deal with null pointers
for the activated window. It is in fact necessary to call it even in
case of a null pointer as it is the only way to get a
ApplicationDeactivated event properly queued.

Without this patch no ApplicationDeactivated event get delivered when
the platform plugin informs the Qt event system of the corresponding user
interaction.

Note: This doesn't apply to qtbase which has the proper behavior, this bug
is Qt4 specific.

Change-Id: I1128d0233a8f85943cfe5fc3c298a43f66b6d4aa
Reviewed-by: Giuseppe D'Angelo <giuseppe.dangelo@kdab.com>