DirectFB: Improvement for QDirectFBPaintEngine::fill()
Improvement for bug fix related to QTBUG-23850 in QDirectFBPaintEngine:
combine the conditions for an accelerated fillrect (5 elements and
rectangle shape) in first if-statement.
The previous check for NULL pointer solves the problem for the crash,
but it simply skips the type check (which caused the crash before) and
the or-condition (shape) became true and leads to an accelerated
fillrect. Otherwise it would fallback to raster.
The original reason for the crash (elemets is null) might be
Q_DISABLE_COPY macro of QVectorPath, especially with the last lines of
the Q_DISABLE_COPY documentation, which could become true on embedded
devices, where DirectFB is often used, with special
toolchains/compilers.

Task-number: QTBUG-23850
Change-Id: I340291fbce24260c7522684305c292e66cfc32af
Reviewed-by: Girish Ramakrishnan <girish.1.ramakrishnan@nokia.com>