Fix dangling pointer issue in QGraphicsItem focus handling
When focused QGraphicsWidget is being hidden, it tries to move the
focus to next focusable widget by calling focusNextPrevChild(true). If
there is no such widget available, the focus is cleared altogether.

The logic used to rely on the return value of focusNextPrevChild to
decide if clearing the focus is necessary. This was incorrect because
the return value may be true also when the focus has actually moved
within the graphics widget in question rather than moving to the next
widget. This is the case at least when GraphicsWebView item is
focused. Thus, the focus could then be left on invisible item which was
a serious issue: elsewhere in the qgraphicsitem.ccp, it is assumed that
no visible item can point to invisible item with its subFocus pointer.
Such pointers will become dangling pointers once the pointed
graphicsitem is deleted, and the process will then crash as soon as the
focus is tried to be moved to another item.

Now, the logic has been modified so that it is explicitely checked
after call to focusNextPrevChild if the item being hidden still has
the focus. If this is the case, the focus is cleared. To be consistent,
the same logic is applied also when disabling a QGraphicsWidget. Having
focus on disabled item is not as severe issue as having it on invisible
item, but it is still wrong.

Task-number: ou1cimx1#995710
Change-Id: Ica32bc381befc3ccaac79fb4cf4d50c5d452fad0
Reviewed-by: Gunnar Sletta <gunnar.sletta@nokia.com>
Reviewed-by: Pasi Pentikäinen <ext-pasi.a.pentikainen@nokia.com>