Eliminate QTreeWidget drag crash
In QTreeViewPrivate::adjustViewOptionsForIndex() wrong index had been
used when referencing to array of viewItems. Variable row is set to the
index of the QModelIndex, however it is not as same as the index in
viewItems[] when there was hidden item in treeWidget. Index of viewItems[]
should be used here. Unit test is added as well.

Change-Id: Ie129cb63445bf1239ef7c5b2bc61b04dd9e81982
Task-Id: QTBUG-25333
Backport of <3be6ca3ee8c2c72f85b26a57538fd329ed811261> from Qt5
Reviewed-by: Girish Ramakrishnan <girish.1.ramakrishnan@nokia.com>