Avoid crash cancelling timer in wrong thread on terminate
When the Symbian timer active object detects that it is cancelled in the
wrong thread, it was issuing a qFatal. But that was an overreaction, as
the situation is not necessarily fatal if the thread in which the timer
lives is about to die. For example if QThread::terminate is used outside
of the thread targeted. Instead a debug message should be issued so that
if there is a crash, the programmer should get information as to why.

The debug message, in this case, is issued with RDebug::Print to avoid
the use of Qt's message output which itself was crashing in partially
finished thread situations.

Task-number: ou1cimx1#1006699
Change-Id: I7d192eb27ba85ba2ef4f80fec3d8479f0df8f90c
Reviewed-by: Juha Kukkonen <ext-juha.kukkonen@nokia.com>
Reviewed-by: Pasi Pentikäinen <ext-pasi.a.pentikainen@nokia.com>
(cherry picked from commit 8ccf7283a1a90d125a774027c6bc3028ab0d29f7)