Fix multiple tablet hires coordinate bug under X11
On X11 if you have multiple tablet input devices, like a Wacom Cintiq and
a Wacom Intuos tablet connected at the same time, one of the tablets will
generate incorrect high resolution coordinates.

The reason seems to be that when translating the XEvent in
qapplication_X11.cpp the code searches for the first tablet device it
finds that supports the type of the current event. Since both tablets
will support this event type, we wind up always finding only one of the
tablets. Specifically, the first tablet that was found when enumerating
tablets during qt_init. Then all calls to QTabletDeviceData::scaleCoord
are made on the same device, so the size of one of the devices is never
used. That device will wind up with incorrect high resolution
coordinates.

When not running under IRIX the XEvent will have a deviceid which we can
use to find the correct device to use for scaling.  This change
implements that fix.

Change-Id: I667a52a729273a9b12880592d0a4cd7ce7f16106
Reviewed-by: Gunnar Sletta <gunnar.sletta@digia.com>
Reviewed-by: Konstantin Ritt <ritt.ks@gmail.com>
Reviewed-by: Lars Knoll <lars.knoll@digia.com>