Fix 'make install' when cross-compiling on Windows
When cross-compiling on a Unix-like environment in a Windows host
we let the default mkspecs be copied to the target directory, instead
of being symlinked.
On Windows, there is an extra-line on the install_mkspecs Makefile target
which also copies (rather than linking) the default mkspecs to
$QT_INSTALL_DIR/mkspecs/default. So in the end this patch removes a duplicate
which was not working, causing the install_mkspecs target to fail.

Change-Id: I44be3a55c11cd7916b4b9ea6a0db1f8d97d28353
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@nokia.com>