QUrl: fix thread safety.
Developers expect const methods on the same QUrl instance to be usable
from multiple threads (including copying and modifying the copy).
The delayed parsing and internal cache of encoded strings break this, however
(and the implicit sharing, for the case of copying).
Protection with a mutex fixes this.

Qt-5.0 doesn't have this issue, since QUrl doesn't do delayed parsing
anymore.

Change-Id: Ie2674e46eb7416b0e753323b673c10f9d3457f6d
Reviewed-by: Marc Mutz <marc.mutz@kdab.com>
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>