Introduce ImprovedSearchHeuristics flag to QLibrary
This commit introduces a new flag to safely introduce the improved
search behaviour in Qt5's QLibrary. If the flag is set then QLibrary
will use the following search heuristic. Otherwise the existing
behaviour will be used.

New search behaviour:

If an absolute path is specified we try that first. Otherwise we first
try the most likely system-specific format (e.g. libfoo.so) on Unix.
This improves performance especially on systems with slow flash devices.

For example, prior to this commit loading the Xcursor library (in the
xcb plugin) results in attempts to dlopen:

"Xcursor"
"Xcursor.so.1"
"libXcursor"
"libXcursor.so.1"

With this commit this is reduced to a single attempt of:

"libXcursor.so.1"

Plugin loading uses absolute paths with QLibrary so there is no
performance penalty for plugins with this commit.

There will be a follow-up commit to make use of the new flag within Qt
to improve performance.

This is a backport of 0026b80cd2a484ad9d685ff5a4f89e6c9815f913

Change-Id: I0dbc83c2909713c01dc687ab8cc3cb0619d1500a
Reviewed-by: Lars Knoll <lars.knoll@nokia.com>