Don't discard QPA mouse events without an associated window or widget
The corresponding window and widget are found by the function later on
if not provided by the plugin, so having a null widget pointer is valid.
This check is unnecessary and breaks integration of QPA plugins not
widget-aware, such as the LinuxInput and TsLib ones.

The bug has already been fixed in Qt5.

Change-Id: I650308ce14e84d42b0c6d64f4f0ae1720728054c
Reviewed-by: Girish Ramakrishnan <girish.1.ramakrishnan@nokia.com>