Fix duplicate mouse event being sent for LinuxInput touch events
This is due to the X and Y messages being received before the TOUCH one,
followed by a SYNC message which then creates the second mouseEvent.
The event should actually be sent on the SYNC message, so remove the
one sent from the TOUCH message handler, just set the posChanged flag.

For more details see:
http://www.kernel.org/doc/Documentation/input/event-codes.txt

Change-Id: I30e71fc3e7da24d8e698467b4d6e0bcd5a64eb37
Reviewed-by: Paul Olav Tvete <paul.tvete@nokia.com>
Reviewed-by: Girish Ramakrishnan <girish.1.ramakrishnan@nokia.com>