Properly handle include cycles
Currently lupdate is just issuing a warning when it detects an
include cycle. Since it's not a full preprocessor, most of those
warnings are false positives, but they need special handling, this
patch provides that.

Now every result is stored as part of a cycle (so in the general
case a cycle of one file). When detecting a cycle, instead instead
of issuing a warning, we record the presence of the newly detected
cycle for later use and merge the other relevant cycles with it.
When retrieving results now it's not for a single file anymore but
for a complete cycle at once (which can turn out to be a cycle of
one file).

This is a backport of the commit done in qttools with revision
3b99677b3376924e347ec4aeb0c363438146bfde

Task-number: QTBUG-6587

Change-Id: Idc79fbac4f69478ccff1464c1f4278d4afbf9d61
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@nokia.com>