Register less types for QtDeclarative/QCoreApplication
In Qt5 QtQml and QtQuick are separated so you have control on the
amount of facilities you need. It is not the case in Qt4
unfortunately so you always pay the price of those registration.

To overcome that, this patch avoids quite some of the type
registration when we detect we're not running a QApplication. In
such case all the GUI related QtQuick 1 runtime is not usable
anyway.

We keep around only a handful of convenience like Timer, Connection,
Component, etc.

Change-Id: Idf5aa935795a224ceb5bc66015282fd094b5686c
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>
Reviewed-by: Chris Adams <christopher.adams@nokia.com>