QGLGlyphCache: Fix texture buffer overrun
The QGLGlyphCache::cacheGlyphs function reallocates a larger texture
when there is no more room to insert a newly rendered glyph. However,
the glyph width used to check whether reallocation is needed is not the
same as the actual glyph image width and may be less. When the glyph
image is then copied into the texture, a buffer overrun may occur.

Task-number: QTBUG-23584
Change-Id: I71d6cd987b7519e5235109c14a5a35e452332417
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@nokia.com>
Reviewed-by: Andy Shaw <andy.shaw@digia.com>
Reviewed-by: Jiang Jiang <jiang.jiang@nokia.com>