Update the state of native menus correctly on Mac
The state of native menus is currently updated in two incompatible ways
on Mac, either by updating the state of native menu items in a native
menu (when setEnabled is called) or the state of the native menu
item containing the native menu (when the menus are initialized and
when a modal dialog is opened/closed). This causes various problems,
e.g. inability to enable a menu that was disabled before menu
initialization or menus getting enabled incorrectly after a modal
dialog is closed. Fix by always setting the state of the native menu
item that contains the native menu.

Also, autoenable is active for the main menu, resulting in unnecessary
menu validation calls, so turn it off.

Task-number: QTBUG-25544
Task-number: QTBUG-26399
Task-number: QTBUG-7538
Change-Id: Iab335af633604e7b3017eaa55464d1b8fdd7b821
Reviewed-by: Morten Johan Sørvig <morten.sorvig@digia.com>