Fixed the QTreeView expansion/collpasing when animated
If you had a QTreeView with expandable items, if you tried to expand and while
the animation was still running you'd try to collpase the node,
the display would be completely broken: the items below that items would
not be visible any more except for a fraction of a second when expanding
or collapsing it again.
The problem is in the fact that when starting an animation the QTreeView
stores the state before animating. And it does that even if an animation
is already running. So the stateBeforeAnimation becomes AnimatingState and
when the animation finishes, AnimatingState is the state that is restored
breaking the painting.
Unit test is included.

qtbase-sha1: 1e97dbaf6ca807397e3ec77a3611763769499d17
Change-Id: I62e16101b70153f78022f6195fd9de6db0cd4878
Reviewed-by: Olivier Goffart <ogoffart@woboq.com>