QWizard/Win: Fix vista helper cache state invalidation
If the cached state is not dirty and the instanceCount is zero, when the
instanceCount is incremented the cached state will not be correct. To
fix this, reset the cached state to dirty if the instanceCount is
incremented from zero.

Change-Id: Ic49eef7f83b1289a939f998817b1b2b5f5a2a45f
(cherry picked from commit e5ebcbc6343a5d5d815cbbd1812f1f5791b7bc88)
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@nokia.com>