Revert "Speed up QTextEngine::resolveAdditionalFormats"
This reverts commit 49fbb6e14af78f751fa114c3cf986aaa607d7408 which
introduced crashes e.g. in Qt Creator. Since the original
change was an optimization, it is reverted in the Qt 4.8 branch
since stability is the main priority there. The crash has been
fixed for the corresponding Qt 5 commit.

It would be an option to recommit the optimization and the
crash fix to a future minor release of Qt 4.

Task-number: QTBUG-27140
Change-Id: I9dc959ea252fe55a22fd90f333ccb335b3ef1311
Reviewed-by: Olivier Goffart <ogoffart@woboq.com>
Reviewed-by: Lars Knoll <lars.knoll@nokia.com>