Blackberry mkspecs: Refine compiler options
stack-protector-strong gives performance benefits over
stack-protector-all and is still checking more than -stack-protector,
so seems to be a good middle way and we want to use it when it is
there.

The -shared option for the compiler (not the linker) prevents a
RIM internal version of qcc from forcing -fPIE, and should not harm
in general when set.

In addition, add a method "compilerSupportsFlag" for Windows as is
present in the Unix configure script.

Original-patch-by: Greg Bentz

(cherry picked from commit 80f6d7862c0e2e41768620d5bd81b0e1d5e3f61f)

Change-Id: I5867fc03dde1ef6e2fbf3747bdb9aaf15518eb6a
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>