Include the signal code in the argument of (dis)connectNotify().
As it turns out, the convention of (dis)connectNotify() is to include
the signal prefix, i.e. '2'.
Therefore add this prefix also when calling these functions from QML.

Also add a unit test confirming that the C++ and QML cases are now
handled the same way.

This patch is not needed in Qt5, as connectNotify() and
disconnectNotify() take a QMetaMethod as a parameter, not a
const char*.

Change-Id: I3add0fc13c60a479949cf3d31218af5fd3f546a2
Reviewed-by: Lars Knoll <lars.knoll@digia.com>