Revert "qmake vcxproj generator: map TARGET_EXT to $(TargetExt)"
This reverts commit 06f0d8a1a9a1d8d46e6f4f3894d76781dfb6638c.

The insane logic that prepends the version number of a lib to
TARGET_EXT instead of appending it to TARGET breaks every vcxproj
with a library version.
Fixing this properly will require much more work and testing on all
supported Windows platforms. Therefore we're reverting the fix for
QTBUG-26782 and mark it as "won't fix" for Qt4.

Task-number: QTBUG-26782
Task-number: QTBUG-27554
Change-Id: I0e3d0f0f3976e9ca958574364ec1b2668e1f66f1
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@digia.com>