QHeaderView - length returns wrong value fix
setSectionHidden called with (logindex, true) will
sometimes not do a (correct) call to
resizeSection(logicalIndex, 0) at once.

However it does execute d->doDelayedResizeSections().
Therefore the section is going to be hidden later.

However it is a problem that the length meanwhile is wrong.
(That is a value that is not the sum of the sections)

This is fixed by execute updates before returning the
length.

This is a backport of SHA 125016ad241125176e5bebab94eebcf50fac20bc

Task-number: QTBUG-14242

Change-Id: Ia1d2f6db3213792b250a6a37942b56554261cd3a
Reviewed-by: Stephen Kelly <stephen.kelly@kdab.com>