don't set QT_BUILD_DIR as a side effect
we would be setting QT_BUILD_DIR to $$[QT_INSTALL_DATA] as a fallback,
which would make qtPrepareTool() believe we were doing a qt build
despite actually trying to build something against an installed qt. this
would become apparent only when using a non-default directory layout, as
is the case in our mac packages.

Change-Id: I4f27e9427222ad3b9bb83fcc625bab19e00db9be
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@digia.com>
Reviewed-by: Teemu Katajisto <teemu.katajisto@digia.com>
Reviewed-by: Joerg Bornemann <joerg.bornemann@digia.com>