fix bug in QLocalSocket::waitForReadyRead on Windows
We must not close the socket, if there's still data in the read buffer.
Also waitForReadyRead must return true, even if the pipe is broken after
we've read data.
QLocalSocket::readData will close the socket after the buffer has been
drained.

This fixes the flakiness of tst_QLocalSocket::threadedConnection.

In Qt5 large portions of this code has been rewritten and this fix does
not apply.

Task-number: QTBUG-27816
Task-number: QTQAINFRA-574
Change-Id: I467340d4dbab11056e6720b145a94a87156cb419
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@digia.com>