Qt 4 - QHeaderView - correct drop on manual QHeaderView sectionmove
This fixes an error in the calculation of the exact drop position
when an user is moving a section in QHeaderView.

Before we compared a mouse-position local to the widget
(pos) with a summed length of sections (posThreshold).

However we need to consider/substract the headers offset to make
the posThreshold comparable to the local mouse position.

This is a backport of SHA d625535728154fc9ca576bf6472c6b5057a96f17

This solves e.g.
Task-number: QTBUG-14814

Change-Id: I8e2c820dc9018fb3d14b71f28aa317417c997a44
Reviewed-by: Stephen Kelly <stephen.kelly@kdab.com>