BlackBerry: Removed filtering of window system events
Only one event type can be passed to the event filter in Qt4. For the
BlackBerry platform this will be BPS events. This should be acceptable,
as there isn't any documentation stating that it should be window system
events for QPA ports.

In Qt5 the implementation is different (it is possible to pass the event
type along with the event itself).

Change-Id: I95550cec8a04a4994479904bc68544742a458aa9
Reviewed-by: Thomas McGuire <thomas.mcguire@kdab.com>