Add a method that allows registration of files to types
There is currently no way in C++ to duplicate the functionality of a qmldir
file in mapping QML files to versioned types in a module. This functionality
would be useful both in cases where a separate qmldir file would be overkill,
and for cases where the type mapping should be generated dynamically.

Since public API is frozen for 4.8, this is being added as private API
for those who need some measure of compatibility with qtquick1 in Qt5.

Change-Id: I28d7898122c5556fcd7cf3476795bcf4bb288ea6
Manual cherry pick of qtquick1/9995a2910d8a5f0317fe3adeb54f838b99ab31a8
Reviewed-by: Matthew Vogt <matthew.vogt@qinetic.com.au>
Reviewed-by: Lars Knoll <lars.knoll@digia.com>