Delay loading implicit import
As a performance improvement to avoid accessing the filesystem
unnecessarily, only import "." implicitly if types cannot be found in the
existing imports. This is not a behavior change for type resolution,
because "." already has the lowest precedence for type resolution.

Change-Id: I5ac2a9fac85559eb96cba93c29d17068fe8171da
Manual cherry-pick of qtquick1/dc96bfd00152e25f007511f64bff7c413f657886
Reviewed-by: Christopher Adams <chris.adams@jollamobile.com>