Update defaultNumberingSystem value for some indic and slovak countries/regions
Qt gets the information about locale digits from the <defaultNumberingSystem>
entries in the CLDR data xm files. Current CLDR data in Qt4.8 is 2.0, which
specifies native digits for most of Indic locales and arabic digits for slovak locale.

This doesn't concern Qt5 that uses the most recent version of CLDR (22.1) which specify Latin digits for the locales in question.

This patch updates the defaultNumberingSystem (to latin) and provides an autotest to compare the returned vs expected digits for the following langagues/regions:

- Gujarati, Hindi, Kannada, Konkani, Malayalam, Marathi, Nepali, Punjabi, Tamil, Telugu (to use latin digits instead of native ones)
- Slovak (to use latin digits instead of arabic ones)

Change-Id: I04364cdb80292f7705b1321377623298a01a9c5b
Reviewed-by: Vladimir Minenko <vminenko@rim.com>
Reviewed-by: Konstantin Ritt <ritt.ks@gmail.com>