Fix QWidget::showFullScreen() on QPA
When showing full screen, we need to enforce that the size of the widget *is*
the size of the screen. When calling QWidget::resize(), there was a chance for
the layout to bypass the screen size, causing the toplevel widget to be
clipped.

This fix is not necessary on Qt5. Qt5 implements this differently and
correctly.

Change-Id: I1f4467dd14310ee20376caba8780507f8d787720
Reviewed-by: Samuel Rødal <samuel.rodal@digia.com>