Try harder in matching the method signature
In cases where the Propertiesdialog has no specific Implementation
for a dbus data type, it creates a regular QLineEdit. In such cases,
qdbusviewer then tried to call the method with the arg set to a string
instead of the real datatype (for example uint64), which usually does not work.

We can be more clever and try to implicitely convert the user input
to the right datatype, which makes the invocation succeed.

Change-Id: I9e16be6e7c5209745f2fa0c32acf08a3c42178f1
(cherry-pick of qttools/0c22eb9d9d1a0b72622fdf16ebf60712336b449e)
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>