Define Q_COMPILER_AUTO_FUNCTION in Qt 4.8.
This definition was missing. It's clear MSVC supports this feature
because it has been in use in qtconcurrentrun.h under an #ifdef
Q_COMPILER_DECLTYPE, which is defined for MSVC.

This is needed in order to re-enable lambdas and other functors in
QtConcurrent after the bug fix for QTBUG-28984.

Task-number: QTBUG-28984
Change-Id: I8c3994096dad120a47ad5f0a7dac6bdbe304fcae
(partly cherry-picked from Qt 5 commit 932034b3df322e7a9dc3d7979584019b46277e50)
Reviewed-by: Olivier Goffart <ogoffart@woboq.com>