Limit the range of the QUrlPrivate::port to -1 to 65535
The internal parser can read values outside this range (and cannot
report an error), but QUrl::port() must not return something outside
that range. The correct solution would be to report an error, like in
Qt 5, but that cannot easily be done.

The rewritten parser in Qt 5 is not affected by this issue.

Task-number: QTBUG-28985
Change-Id: I3cf595384f14272197dcfb85943213c8f8ddeba0
Reviewed-by: David Faure (KDE) <faure@kde.org>