set pkg-config *_location directly to install dir
qmake tries to replace all occurrences of the install prefix with a
literal ${prefix}. of course this would mess up build paths which happen
to live under the prefix (this is untypical, but still), which would
break the sed magic in the subsequent install step. as qmake itself has
no use for the .pc files, it's ok to target them directly for the
install dir. we do the same with the include and lib dirs already.

qt5 is already doing the right thing.

Task-number: QTBUG-28807
Change-Id: I82c17284d1dbac4a96845db2ddec9680e251e7a5
Reviewed-by: Davide Pesavento <davidepesa@gmail.com>
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@digia.com>