fix illegal reference to a destroyed variable
The char array 'dash_o' is an automatic variable. The string "-o" is copied into this array.
Later the address of the array dash_o is assigned twice to lpargs[++i].
After leaving the block, the array dash_o is gone and lpargs[...] contains an illegal reference.

This was discovered in a release mode when compiled with gcc version 4.7.2

Patch doesn't apply for Qt5 where 'lpr' support has been removed.

Change-Id: I4f99badfa380ad3b29893a350f0d699bfb934c68
Reviewed-by: Olivier Goffart <ogoffart@woboq.com>