BlackBerry bearer plugin: check whether device is online several times
This is supposed to workaround a race condition in the underlying
netstatus API: Sometimes we get an event that the Wifi interface
changed, but it is not up, e.g. no gateway (yet). In that case we need
to check back (currently: 300 ms) whether the interface has come
up or not.
This commit can be reverted again once the race condition in the
netstatus API has been resolved.

Task-number: QTBUG-29421
(cherry picked from commit 3dc47622a469ded1c99397fdedc2053b73d57189)

Change-Id: I252c39411340fe3f6af8440a8f21482ca9b6e62d
Reviewed-by: Rafael Roquetto <rafael.roquetto@kdab.com>