Fix text issue with Hiragana characters when stylesheet is applied.
When typing into a QTextEdit or QLineEdit with a Japanese IME on
Windows, some characters have both a black foreground and background
and hence cannot be seen. This patch fixes the issue by setting the
appropriate colors for the text in these scenarios.

The change also fixes a similar situation with QSyntaxHighlighter and
Hiragana, where the background and foreground colors were the same.
Now it will use underlining to highlight the text (like WordPad does,
for example).

Task-number: QTBUG-29442
Change-Id: I364d327a8f6573a73b227b79cc5f570cd02611c9
Reviewed-by: Andy Shaw <andy.shaw@digia.com>
Reviewed-by: Titta Heikkala <titta.heikkala@digia.com>
Reviewed-by: David Faure (KDE) <faure@kde.org>
Reviewed-by: Mitch Curtis <mitch.curtis@digia.com>