QNX: Don't let eglMakeCurrent() fail when using two contexts.
When calling eglMakeCurrent(), egl config and window format
must match, otherwise it fails with EGL_BAD_MATCH.

egl config and window format were static variables, this could lead
to a context setting the window format in the ctor but, at the
time of creating a surface, another context had changed the window
format.

elg config and window format are now member variables. This is the
same approach taken in Qt5.

Change-Id: I4fa0bc59273b856c1eaf43fafd15b551363f2fe2
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>