Fix path separators in archives created by QZipWriter
It was not possible to extract data from the archive on OSX
which was created on the Windows platform because of wrong
separators. Archive was created on Windows via QZipWriter and
opened on OSX with QZipReader. It consisted of a lots directories
and subdirectories with files. The solution is to use '/' separator
for internal representation.

Cherry picked from qtbase eff6dbb306956f099374e5a0c5942b9e91a81e71

Change-Id: I2180282d235dc998309fc4f7b67f5c0181b78dda
Reviewed-by: Vasiliy Sorokin <sorokin.vasiliy@gmail.com>
Reviewed-by: Konstantin Ritt <ritt.ks@gmail.com>
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@digia.com>