Fix Mac OS X crc/dwarf2 test for non-default Xcode setup.
The crc.test and dwarf2.test scripts are now using the
SDK and arch settings passed to it. If you install Xcode
without the "UNIX Development" option, this is essential
to let the compiler find standard headers and let the test
program compile successfully.

In addition, let configure pass the SDK and arch settings
given on the command line to the scripts.

These scripts are no longer present in Qt 5.

Change-Id: I78c0b3594b91827a6618a9a152949e9201ca64c8
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@digia.com>
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@digia.com>