unmess CONFIG/QT_CONFIG static/shared
bring it to the state of 5.0.2: both QT_CONFIG and CONFIG in qconfig.pri
contain the respective flag, so it is a) visible how qt was built and b)
user code has a consistent default.

before, on windows we had it only in qconfig.pri's CONFIG (so it was
harder to find the qt config), and on unix we had it only in
.qmake.cache's CONFIG (so it was impossible to find the qt config in a
clean way, and there was no default (other than some arbitrary mess
inside qmake itself)).

Change-Id: Iad06de8c4435cec3e1407babde037ac96cb5bf66
Reviewed-by: Joerg Bornemann <joerg.bornemann@digia.com>