Windows: Fix the last file dialog bottleneck.
Went from taking 30 seconds to 2 seconds, on a SDCard with 10k files.

Windows file dialog does not resolve NTFS symlinks,
it just shows an empty icon, and the link name, not the target.

This allows for a big performance gain by reducing the number of
calls to GetFileAttributesEx() by checking the extension directly.

This also fixes the problems with the native file dialog, which
for some reason, is creating a QFileSystemModel too.

Task-number: QTBUG-13182
(cherry picked from commit 51f00deffac49c4277425837f0132b0c721bb689)

Change-Id: If7bd63d68e1870c5e48907ae79f4c4bdc6972b00
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@digia.com>