QNetworkConfigurationManager: check whether app is shutting down
... before trying to update the configurations. Before, we would check
whether the pointer to the QNetworkConfigurationManagerPrivate instance
was 0 and in that case construct it. This would mean that this code path
was taken "at app shutdown", i.e. when the qAddPostRoutine had already
been called but the other statics were still accessed.

Note: This is not thread safe, but neither is the rest of the code;
making it thread-safe would require additional changes.

Task-number: QTBUG-30585
(backport of commit cc028650e590ac49293558099b3fb87720e9793b)

Change-Id: Ifaae4d758a10b44bc7f838cdc0d3a1129ed63228
Reviewed-by: Stephen Kelly <stephen.kelly@kdab.com>