Fix Qt WebKit build for non-default Xcode setup / Xcode 4.3+
There's no need to test the full path, the actual name of the SDK will do.
This fixes the build in cases where Xcode is not installed in /Developer
or when building with Xcode 4.3 or higher (where Xcode is a bundle), because
in those cases the SDK's full path does not start with /Developer.

Change-Id: I598ba2a5ebaae813f576dc72102d05fe8804a86f
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@digia.com>