[BB10-internal] SSL internals: fix memory corruption using QSslConfigurationPrivate
We are passing a QSslConfigurationPrivate that is allocated on the stack
(in QSslSocketBackendPrivate::initSslContext()) to
QSslConfiguration::QSslConfiguration(QSslConfigurationPrivate *dd).
When the SSL context is destroyed, this object is not there any more.
So now we create a deep copy of the configuration like we do in
QSslSocket::sslConfiguration().

Task-number: QTBUG-30648
(backport of commit 3a43aff9deb4af0479914a26d68fb98d313369b6)
Signed-off-by: Peter Hartmann <phartmann@blackberry.com>

Change-Id: I0a39b2bc485ce5a3528b72e6e47c3bd124963b3b