Revert "JSCore build fix for WEC7"
I's reverted, because there seems to be at least two
implementations of the stl in WEC7. One for each BSP SDK
and one shipped with the compiler, which is in this
case Visual Studio 2008. The Compiler has a stl with std::min
etc. Visual Studio GUI also uses the compiler version of
the stl first. It seems to be that the environment was
setup wrongly. Checksdk also uses the right order.

The same patch is reverted for Qt5 as well:
https://codereview.qt-project.org/25482

This reverts commit 1fba05343eb6fea40f6617a62e31e10a0e918ec2

Change-Id: I956cc6f0669c67a094a2af464f8aa719a137a7a5
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@digia.com>
Reviewed-by: Björn Breitmeyer <bjoern.breitmeyer@kdab.com>
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@digia.com>