Special-case the forward declaration of STL types for libc++.
This is somewhat of a follow-up to 5210d47aa66214e3cb16f394d0510a91f770c1b1.
libc++ declares the STL types in an inline namespace within the "std" one if
clang is used.

If one includes a header such as <iterator> before a Qt one and builds with
QT_NO_STL, the following ends up happening:

  // <iterator>
  namespace std {
      inline namespace __1 {
          struct random_access_iterator ... ;
      }
  }

  // qiterator.h
  #ifdef QT_NO_STL
  namespace std {
     struct random_access_iterator;
  }
  #endif

qiterator.h's declaration shadows the original one, and the compiler
complains random_access_iterator lacks an actual declaration.

Solve this by checking for libc++ and forward-declaring the iterator types
we need within the same inline namespace.

Not backported from qtbase because QT_NO_STL does not exist in Qt 5.

Change-Id: I6742d540f6538a30aa060a4447c288cfb9cd781d
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>