Respect specified minimum height for menuBar also if it has HFW.
If the menu bar is subject to height for width (HFW) we should of
course respect that, but in addition we should ensure that the HFW is
within the minimum and maximum height. This also is consistent with how
QGridLayout calculates the effective minimum row height.

This fixes a regression because change 4780f94e391b5e881497c5228661dead
turned QTabWidget into a proper height-for-width citizen, and when
setting a QTabWidget as a menuwidget, the buggy codepath for HFW was
suddenly hit in menuBarHeightForWidth().

Task-number: QTBUG-31057

Reviewed-by: Paul Olav Tvete <paul.tvete@digia.com>
(cherry picked from qtbase/27e690e163408dec1e9c56ffc07131354d4b2c8a)

Change-Id: I915d37c69152c1804925000303e82a3fae5a9a47
Reviewed-by: Paul Olav Tvete <paul.tvete@digia.com>