QGtkStyle: add getIconThemeName() helper function
Add QGtkStyle::getIconThemeName() static helper function
(which uses GtkSettings to retrieve the theme name) and use it in
QGuiPlatformPlugin::systemIconThemeName() when desktop is GNOME.

In modern distributions like Debian and Ubuntu, the
/desktop/gnome/interface GConf schema is no longer installed,
so the old method for retrieving icon theme name no longer works.

Qt 5 already uses GtkSettings for this thing, although the code
is in a different place.

Change-Id: Ib13d4809b55dd55dc1fa0bc2c4b1552f13feb030
Reviewed-by: J-P Nurmi <jpnurmi@digia.com>
Reviewed-by: Jens Bache-Wiig <jens.bache-wiig@digia.com>